+++ tests/python/fake_bpy_module_test/fake_bpy_module_test/transformer_test_data/bpy_module_tweaker/expect/add_bpy_app_handlers_functions_data_types_transformed.xml
@@ -1,6 +1,6 @@
 <document source="<string>">
     <source-filename>
-        basic.rst
+        add_bpy_app_handlers_functions_data_types.rst
     <module>
         <name>
             bpy.app.handlers
