+++ tests/python/fake_bpy_module_test/fake_bpy_module_test/transformer_test_data/bpy_module_tweaker/expect/add_bpy_ops_override_parameters.xml
@@ -1,6 +1,6 @@
 <document source="<string>">
     <source-filename>
-        basic.rst
+        add_bpy_ops_override_parameters.rst
     <module>
         <name>
             bpy.ops.pose
