+++ tests/python/fake_bpy_module_test/fake_bpy_module_test/transformer_test_data/bpy_module_tweaker/expect/make_bpy_prop_functions_arguments_kwonlyargs.xml
@@ -1,6 +1,6 @@
 <document source="<string>">
     <source-filename>
-        basic.rst
+        make_bpy_prop_functions_arguments_kwonlyargs.rst
     <module>
         <name>
             bpy.props
@@ -11,7 +11,7 @@
         <description>
             function_1 description
         <argument-list>
-            <argument argument_type="kwonlyarg">
+            <argument argument_type="arg">
                 <name>
                     arg_1
                 <description>
@@ -20,7 +20,7 @@
                 <data-type-list>
                     <data-type>
                         function_1 arg_1 type
-            <argument argument_type="kwonlyarg">
+            <argument argument_type="arg">
                 <name>
                     arg_2
                 <description>
