+++ tests/python/fake_bpy_module_test/fake_bpy_module_test/transformer_test_data/bpy_module_tweaker/expect/rebase_bpy_types_class_base_class.xml
@@ -1,6 +1,6 @@
 <document source="<string>">
     <source-filename>
-        basic.rst
+        rebase_bpy_types_class_base_class.rst
     <module>
         <name>
             bpy.types
