+++ tests/python/fake_bpy_module_test/fake_bpy_module_test/transformer_test_data/bpy_module_tweaker/input/make_bpy_prop_functions_arguments_kwonlyargs.rst
@@ -1,6 +1,6 @@
 .. module:: bpy.props
 
-.. fucnction:: function_1(arg_1, arg_2=5.0)
+.. function:: function_1(arg_1, arg_2=5.0)
 
    function_1 description
