+++ tests/python/fake_bpy_module_test/fake_bpy_module_test/transformer_test_data/bpy_module_tweaker/input/rebase_bpy_types_class_base_class.rst
